Done: 예약 완료/취소 공통화
